Updated layer export
The labels for the planet names are no longer grouped onto one layer. Please update to bring the names of the planetary objects all onto 1 layer.
I noticed that the outer orbit for Neptune seems to be duplicated. Make sure that only one elliptal path is created for each planetary orbit.