Create index.html
Cursor: Can you please make me the index.html file and put it in this project? I'd like to include all of the .html projects in this project as published links. Also include an AI generated description of each project. I'd like this index.html page to regenerate when new .html projects get added to this folder.